separate identify tests
